Correct class names.
